services/textAnalytics/requirements.txt
flask
rake-nltk
requests
python-dotenv
azure-ai-textanalytics --pre
